chore: break `lsp` down into more modules
